requirements.txt
aiohttp
aiosqlite
asyncio
discord.py
Scrapy==2.7.0
